requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9

openai==1.43.0

pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0

boto3==1.34.162
sqlalchemy==2.0.32
psycopg2-binary==2.9.9

pytest==8.3.2
